Load agent from isolated class loader
