+++ apm-agent-common/src/main/java/co/elastic/apm/agent/common/JvmRuntimeInfo.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.premain;
+package co.elastic.apm.agent.common;
 
 import javax.annotation.Nullable;
