+++ apm-agent-common/src/test/resources/test.txt
@@ -1,0 +1,1 @@
+I'm here to help test resource extraction
