+++ apm-agent-core/src/main/java/co/elastic/apm/agent/bci/IndyBootstrap.java
@@ -21,8 +21,7 @@ package co.elastic.apm.agent.bci;
 import co.elastic.apm.agent.bci.classloading.ExternalPluginClassLoader;
 import co.elastic.apm.agent.bci.classloading.IndyPluginClassLoader;
 import co.elastic.apm.agent.bci.classloading.LookupExposer;
-import co.elastic.apm.agent.premain.JavaVersionBootstrapCheck;
-import co.elastic.apm.agent.premain.JvmRuntimeInfo;
+import co.elastic.apm.agent.common.JvmRuntimeInfo;
 import co.elastic.apm.agent.sdk.state.GlobalState;
 import co.elastic.apm.agent.util.PackageScanner;
 import net.bytebuddy.asm.Advice;
@@ -32,7 +31,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.stagemonitor.configuration.ConfigurationOptionProvider;
 import org.stagemonitor.util.IOUtils;
-import sun.misc.Unsafe;
 
 import javax.annotation.Nullable;
 import java.io.File;
@@ -155,7 +153,7 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
  *     <li>
  *         The {@code INVOKEDYNAMIC} support of early Java 7 versions is not reliable.
  *         That's why we disable the agent on them.
- *         See also {@link JavaVersionBootstrapCheck}
+ *         See also {@code co.elastic.apm.agent.premain.JavaVersionBootstrapCheck}
  *     </li>
  *     <li>
  *         There are some things to watch out for when writing plugins,
@@ -172,7 +170,7 @@ public class IndyBootstrap {
      */
     private static final String INDY_BOOTSTRAP_CLASS_NAME = "java.lang.IndyBootstrapDispatcher";
     /**
-     * Needs to be loaded from the bootstrap CL because it uses {@link Unsafe}.
+     * Needs to be loaded from the bootstrap CL because it uses {@code sun.misc.Unsafe}.
      * In addition, needs to be loaded explicitly by name only when running on Java 9, because compiled with Java 9
      */
     private static final String INDY_BOOTSTRAP_MODULE_SETTER_CLASS_NAME = "co.elastic.apm.agent.bci.IndyBootstrapDispatcherModuleSetter";
@@ -224,7 +222,10 @@ public class IndyBootstrap {
         try {
             indyBootstrapDispatcherClass = Class.forName(INDY_BOOTSTRAP_CLASS_NAME, false, null);
         } catch (ClassNotFoundException e) {
-            byte[] bootstrapClass = IOUtils.readToBytes(ClassLoader.getSystemClassLoader().getResourceAsStream(INDY_BOOTSTRAP_RESOURCE));
+            byte[] bootstrapClass = IOUtils.readToBytes(ElasticApmAgent.getAgentClassLoader().getResourceAsStream(INDY_BOOTSTRAP_RESOURCE));
+            if (bootstrapClass == null || bootstrapClass.length == 0) {
+                throw new IllegalStateException("Could not locate " + INDY_BOOTSTRAP_RESOURCE);
+            }
             ClassInjector.UsingUnsafe.ofBootLoader().injectRaw(Collections.singletonMap(INDY_BOOTSTRAP_CLASS_NAME, bootstrapClass));
             indyBootstrapDispatcherClass = Class.forName(INDY_BOOTSTRAP_CLASS_NAME, false, null);
         }
