+++ apm-agent-core/src/main/java/co/elastic/apm/agent/report/ApmServerReporter.java
@@ -23,7 +23,7 @@ import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.impl.transaction.Transaction;
 import co.elastic.apm.agent.report.disruptor.ExponentionallyIncreasingSleepingWaitStrategy;
 import co.elastic.apm.agent.util.MathUtils;
-import co.elastic.apm.agent.premain.ThreadUtils;
+import co.elastic.apm.agent.common.ThreadUtils;
 import com.dslplatform.json.JsonWriter;
 import com.lmax.disruptor.EventFactory;
 import com.lmax.disruptor.EventTranslator;
