+++ apm-agent-core/src/main/java/co/elastic/apm/agent/util/ExecutorUtils.java
@@ -18,7 +18,7 @@
  */
 package co.elastic.apm.agent.util;
 
-import co.elastic.apm.agent.premain.ThreadUtils;
+import co.elastic.apm.agent.common.ThreadUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
