+++ apm-agent-core/src/test/java/co/elastic/apm/agent/util/ThreadUtilsTest.java
@@ -18,7 +18,7 @@
  */
 package co.elastic.apm.agent.util;
 
-import co.elastic.apm.agent.premain.ThreadUtils;
+import co.elastic.apm.agent.common.ThreadUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.Test;
