+++ apm-agent-plugins/apm-java-concurrent-plugin/src/test/java/co/elastic/apm/agent/concurrent/AsyncTraceMethodInstrumentationTest.java
@@ -69,6 +69,8 @@ class AsyncTraceMethodInstrumentationTest {
 
         tracer = mockInstrumentationSetup.getTracer();
         ElasticApmAgent.initInstrumentation(tracer, ByteBuddyAgent.install());
+        new TestAsyncTraceMethodsClass().invokeAsync();
+        reporter.reset();
     }
 
     @AfterEach
